Make a function `get_logger` which replaces the package name of the logger with the predefined `anki_convo` package name
