+++ src/anki_convo/__about__.py
@@ -2,3 +2,4 @@
 #
 # SPDX-License-Identifier: GPL-3.0-or-later
 __version__ = "0.0.1"
+NAME = "anki_convo"
