+++ src/anki_convo/__init__.py
@@ -2,8 +2,8 @@
 #
 # SPDX-License-Identifier: GPL-3.0-or-later
 from . import _paths  # noqa: I001, F401
+from .__about__ import NAME, __version__  # noqa: F401
 
-import logging.config
 import os
 
 from .env import load_env
@@ -12,11 +12,11 @@ from .addition_filter import init_addition_filter
 from .card_count_view import init_card_count_view
 from .field_filter import init_field_filter
 from .llm_filter import init_llm_filter
-from .logging_config import setup_logging
+from .logging import setup_logging, get_logger
 
 
 setup_logging()
-logger = logging.getLogger(__name__)
+logger = get_logger(__name__)
 
 
 def init_package():
